Gps precision squash
* Add precision GPS types to telemetry IDs

* Update telemetry parsing to support mixed types

Telemetry parsing now supports mixed type precision for lat/long gps values

* rework bit shifting to be correct

* Update SerialParser.java

* adjust getTelemetry call to handle double instead of float

* Increased the allowed range for settings to 72

* Convert telemetry float values to doubles across the code base

convert telemetry focused float variables and structures to double variables and structures. This probably breaks the graph and the waypoint panel in some way... Testing will bear it out.

* Reworked precision telemetry value parsing logic

...

* Update SerialParser.java

updates for ground settings

Update groundSettings.xml

Update groundSettings.xml

+++ src/remote/SettingList.java
@@ -18,36 +18,42 @@ public class SettingList {
         context = cxt;
         loadSettingData();
     }
+    
     public int size() {
         return settingData.size();
     }
+    
     public Setting get(int id) {
         return settingData.get(id);
     }
-    public void pushSetting(int id, float val) {
+    
+    public void pushSetting(int id, double val) {
         if(id < 0 || id >= settingData.size()) return;
         updateSettingVal(id, val);
         pushSetting(id);
     }
+    
     public void pushSetting(int id) {
         Message msg = Message.setSetting((byte)id, settingData.get(id).getVal());
         context.sender.sendMessage(msg);
     }
-    public void updateSettingVal(int id, float val) {
+    
+    public void updateSettingVal(int id, double val) {
         if(id < 0 || id >= settingData.size()) return;
         settingData.get(id).setVal(val);
     }
-
-    private float getFloat(XMLStreamReader reader, String label) {
+    
+    private double getDouble(XMLStreamReader reader, String label) {
         String raw = reader.getAttributeValue(null,label);
         if(raw == null) return 0.0f;
         //strip whitespace
         raw = raw.replaceAll("\\s","");
         //parse
-        if(raw.equals("+inf"))       return Float.POSITIVE_INFINITY;
-        else if (raw.equals("-inf")) return Float.NEGATIVE_INFINITY;
-        else                         return Float.valueOf(raw);
+        if(raw.equals("+inf"))       return Double.POSITIVE_INFINITY;
+        else if (raw.equals("-inf")) return Double.NEGATIVE_INFINITY;
+        else                         return Double.valueOf(raw);
     }
+    
     private void loadSettingData() {
         settingData.clear();
         for(int i=0; i<Serial.MAX_SETTINGS; i++) {
@@ -74,9 +80,9 @@ public class SettingList {
                         }
                         tmp = settingData.get(index);
                         tmp.name = reader.getAttributeValue(null,"name");
-                        tmp.min  = getFloat(reader,"min");
-                        tmp.max  = getFloat(reader,"max");
-                        tmp.def  = getFloat(reader,"def");
+                        tmp.min  = getDouble(reader,"min");
+                        tmp.max  = getDouble(reader,"max");
+                        tmp.def  = getDouble(reader,"def");
                         break;
                     case XMLStreamConstants.CHARACTERS:
                         if(tmp != null)
